connect api resonse to dom relates #16
