+++ public/js/script.js
@@ -10,9 +10,8 @@ sendButton.addEventListener('click', function(e) {
   // get input
   var inputField = document.getElementById('user-input')
   var userInput = inputField.value;
-  console.log(userInput);
   // Send input to server
-  sendInput(userInput);
+  sendInput(userInput, renderBotResponse);
   // add message bubble
   var chatLi = document.createElement('li');
   chatLi.id = "user-speech-bubble";
@@ -31,16 +30,30 @@ sendButton.addEventListener('click', function(e) {
   // e.dispatchEvent();
 });
 
-function sendInput(input) {
+function sendInput(input, cb) {
   var encoded = "message=" + encodeURI(input);
   var xhr = new XMLHttpRequest();
   xhr.onreadystatechange = function () {
-    if (xhr.readyState === 4 && xhr.statusCode === 200) {
+    if (xhr.readyState === 4 && xhr.status === 200) {
       var parsed = JSON.parse(xhr.responseText)
-      console.log(parsed);
+      cb(parsed)
     }
   }
   xhr.open("POST","/send-message", true)
   xhr.setRequestHeader("Content-Type", "application/x-www-form-urlencoded");
   xhr.send(encoded)
+}
+
+function renderBotResponse(resJSON) {
+  thinkingDiv.textContent = "";
+  var botLi = document.createElement('li');
+  botLi.id = 'bot-speech-bubble';
+  botLi.textContent = resJSON.fulfillmentText;
+
+  var pText = document.createElement('p');
+  pText.textContent = 'Bot';
+  pText.classList.add('name');
+  botLi.appendChild(pText);
+  chatUl.appendChild(botLi);
+
 }
